Added a SConstruct to build the extension.
Fixed a number of things in spidermonkey.pyx so that it compiles and passes its unit tests.  I assume that all of these are a result of Pyrex or SpiderMonkey changing over the years since this code was originally written.

  * Fixed a bug re: the use of bool() in Pyrex.

  * Changed the signature of JS_GetClass().

  * Changed the 'spare' property of the JSClass struct to the 'reserveSlots' property.

I also made python-spidermonkey deal gracefully with what could be a bug in SpiderMonkey itself, whereby the property getter for a global object is asked for the value of undefined (JSVAL_VOID).